requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
agno>=2.0.0
python-dotenv>=1.0.1
openai>=1.12.0
groq>=0.4.2
pypdf>=4.0.1
sqlalchemy>=2.0.0
pyinstaller>=6.3.0
